Error listener for autoUpdater and fix for preferences file being unlinked on app launch.

+++ resources/functions/handler.js
@@ -3,7 +3,7 @@ const SentryInit = require("./init").SentryInit;
 SentryInit()
 const {LoadOneTimeFiles, LoadFiles} = require('./load');
 const {join, resolve} = require('path');
-const {unlinkSync} = require('fs');
+const fs = require('fs');
 const rimraf = require('rimraf');
 
 
@@ -73,11 +73,18 @@ const handler = {
 
     VersionHandler: function () {
         if (!app.preferences.value('storedVersion') || app.preferences.value('storedVersion') === undefined || app.preferences.value('storedVersion') !== app.getVersion()) {
-            rimraf(resolve(app.getPath('userData'), 'Cache'), [], () => {
-                console.warn(`[VersionHandler] Outdated / No Version Store Found. Clearing Application Cache. ('${resolve(app.getPath('userData'), 'Cache')}')`)
-            })
-            unlinkSync(resolve(app.getPath('userData'), 'preferences.json'))
-            console.warn(`[VersionHandler] 'preferences.json' deleted.`)
+
+            if (fs.existsSync(resolve(app.getPath('userData'), 'Cache'))) {
+                rimraf(resolve(app.getPath('userData'), 'Cache'), [], () => {
+                    console.warn(`[VersionHandler] Outdated / No Version Store Found. Clearing Application Cache. ('${resolve(app.getPath('userData'), 'Cache')}')`)
+                })
+            }
+
+            if (fs.existsSync(resolve(app.getPath('userData'), 'preferences.json'))) {
+                fs.truncate(resolve(app.getPath('userData'), 'preferences.json'), 0, function() {
+                    console.warn(`[VersionHandler] Outdated / No Version Store Found. Clearing Preferences File. ('${resolve(app.getPath('userData'), 'preferences.json')}')`)
+                });
+            }
         }
     },
